1.0.56
Finish Paper
Fix Bug
Update Lang System

+++ application/models/Submit_model.php
@@ -239,6 +239,7 @@ class Submit_model extends CI_Model {
         //$this->db->where("conf_id",$conf_id);
         $this->db->where("fid",$fid);
         $this->db->where("paper_id",$paper_id);
+        if( $file->file_type !="F" || $file->file_type !="O" ) return false;
         if( $this->db->delete('paper_file') ){
             // $this->conf->add_log("submit","del_file",$conf_id,array("paper_id"=>$paper_id,"fid"=>$fid));
             return true;
@@ -248,7 +249,6 @@ class Submit_model extends CI_Model {
     }
 
     function get_otherfile($paper_id){
-        $this->db->select('*');
         $this->db->from('paper_file');
         $this->db->where('file_type', "F");
         $this->db->where('paper_id', $paper_id);
@@ -257,7 +257,6 @@ class Submit_model extends CI_Model {
     }
 
     function get_otherfiles($paper_id){
-        $this->db->select('*');
         $this->db->from('paper_file');
         $this->db->where('file_type', "O");
         $this->db->where('paper_id', $paper_id);
@@ -823,11 +822,11 @@ class Submit_model extends CI_Model {
     function sendmail_submit_success($paper_id,$conf_id){
         $paper = $this->mail_get_paper($conf_id,$paper_id);
         $author_name   = preg_match("/[\x{4e00}-\x{9fa5}]/u", $paper->user_last_name)?$paper->user_last_name.$paper->user_first_name:$paper->user_first_name." ".$paper->user_last_name;
-        $author        = is_null($paper->user_login)?'<a href="'.base_url("user/signup").'">未註冊帳號,前往註冊帳號 / Unregistered Account,Go To Signup Account</a> )':$paper->user_login;
+        $author        = is_null($paper->user_login)?'<a href="'.site_url("user/signup").'">未註冊帳號,前往註冊帳號 / Unregistered Account,Go To Signup Account</a> )':$paper->user_login;
         $paper_title   = $paper->sub_title;
         $paper_summary = $paper->sub_summary;
         $conf_name = $paper->conf_name;
-        $conf_link = base_url($conf_id);
+        $conf_link = site_url($conf_id);
         $submit_link = get_url("submit",$conf_id);
         $user_email = $paper->user_email;
         $conf_email =  $paper->conf_email;
@@ -846,5 +845,56 @@ class Submit_model extends CI_Model {
         $this->email->message($mail_content);
         
         $this->email->send();
+    }
+
+    function get_finishfile($paper_id){
+        $this->db->from('paper_file');
+        $this->db->where('file_type', "FF");
+        $this->db->where('paper_id', $paper_id);
+        $query = $this->db->get();
+        return $query->row();
+    }
+
+    function get_finishother($paper_id){
+        $this->db->from('paper_file');
+        $this->db->where('file_type', "FO");
+        $this->db->where('paper_id', $paper_id);
+        $query = $this->db->get();
+        return $query->result();
+    }
+
+    function del_finishfile($conf_id,$paper_id,$fid){ // $fid array
+        $this->db->from('paper_file');
+        $this->db->where("paper_id",$paper_id);
+        $this->db->where_in("fid",$fid);
+        $this->db->where_in("file_type",array("FF","FO"));
+        $query = $this->db->get();
+        $files = $query->result();
+        foreach ($files as $key => $file) {
+            delete_files($this->conf->get_paperdir($conf_id).$file->file_system);
+        }
+        $this->db->where_in("fid",$fid);
+        $this->db->where_in("file_type",array("FF","FO"));
+        if( $this->db->delete('paper_file') ){
+            return true;
+        }
+        return false;
+    }
+
+    function paper_to_finish($conf_id,$paper_id){
+        $paper = array(
+            "sub_status"=>5,
+            "sub_time"=>time()
+        );
+        $this->db->where("conf_id",$conf_id);
+        $this->db->where("sub_id",$paper_id);
+        if( $this->db->update('paper', $paper) ){
+            // $this->sendmail_submit_success($paper_id,$conf_id);
+            // $this->topic->notice_editor($conf_id,$paper_id);
+            // $this->conf->add_log("submit","paper_to_review",$conf_id,$paper);
+            return true;
+        }else{
+            return false;
+        }
     }
 }
